feat(.github/actions): add optional `setup-node` in update-axe-core action (#208)
This patch adds optional input `should-setup-node` which defaults to
`true` to maintain existing behavior. Some repos already setup node in a
previous step in the action with a particular version there is no need
to re-setup node as it could checkout with a previous version of Node
and/or waste minutes.

No QA Required