feat(action): action to add label of "Closed: {Date}" and remove old labels starting with "Closed:" (#221)
Closes: https://github.com/dequelabs/results-team/issues/66

Adds an action that will add a label of "Closed: yyyy-mm-dd" and remove
any other labels starting with "Closed:" from the issue on issue
closing.

QA Notes: Visit https://github.com/dequelabs/tommy-testing, create an
issue, close it, verify the label was added, update the label to be a
date different than current date, reopen, reclose it, verify the Closed:
is set back to the current date.