ci(action): change hardcoded values to inputs of the release-candidate GHAs (#224)
This pull request enhances the flexibility and configurability of two
internal GitHub Actions: `create-release-candidate-v1` and
`release-candidate-ready-for-qa-v1`. The main improvements are the
addition of new input parameters to support project board automation,
customizable assignees and labels, and more robust conditional logic for
step execution.
These new inputs are optional, and it skips related steps if they are
not provided.

Closes: results-team
[#77](https://github.com/dequelabs/results-team/issues/77)