chore: fix dependabot to update all actions action.yml (#249)
According to
https://docs.github.com/en/code-security/dependabot/working-with-dependabot/dependabot-options-reference#directories-or-directory--,
dependabot is not updating any of our actions in the .github/actions
directory as it only looks for the `action.yml` file in the root dir.
Depending how github defines the "root dir" this could or could not
work. Testing to see if this syntax will allow dependabot to update all
the actions `action.yml` files.

Temporarily took away the interval setting so we could see if thins
works without waiting a month.

No QA required